Create SQL_PHP_WP_CLI_research.md
Ok thank you very much!

So we said that a CSV import is not possible because WP All Import the best plugin for do it, but also all the others are paid only. And I searching for free method.

Translate WXR file is complicated can incur in errors becuase of specific format of the XML file and also a lot of <![CDATA[ ... ]]>. That still can be an option with a well instructed method for translate it!

I thinking now about another method.... can be a SQL export of Database, and after a translate of this DB data and a import directly with SQL?
Or by using wordpress functions directly?

Let's think about it, make me question if you not understanding or need more information, I need to find an automated way for translate this WPML spanish webiste into English, German, Italian, French!